# openapi.json
{
  "components": {
    "responses": {
...
      "Error401": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorBody"
            }
          }
        },
        "description": "Authentication is required or the supplied credential is invalid."
      },
      "Error403": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorBody"
            }
          }
        },
        "description": "The authenticated principal is not permitted to perform this operation."
      },
      "Error404": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorBody"
            }
          }
        },
        "description": "The requested tenant or resource was not found."
      },
      "Error409": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorBody"
            }
          }
        },
        "description": "The request conflicts with the current resource or integration state."
      },
      "Error413": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorBody"
            }
          }
        },
        "description": "The request declares content larger than the configured limit."
      },
      "Error422": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorBody"
            }
          }
        },
        "description": "The request did not pass structural or semantic validation."
      },
...
    },
    "schemas": {
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
...
      "MediaResponse": {
        "properties": {
          "alt_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alt Text"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "mime_type": {
            "title": "Mime Type",
            "type": "string"
          },
          "original_filename": {
            "title": "Original Filename",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "visibility": {
            "enum": [
              "PUBLIC",
              "PRIVATE"
            ],
            "title": "Visibility",
            "type": "string"
          },
          "width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          }
        },
        "required": [
          "id",
          "original_filename",
          "mime_type",
          "size_bytes",
          "width",
          "height",
          "alt_text",
          "status",
          "visibility"
        ],
        "title": "MediaResponse",
        "type": "object"
      },
...
      "UploadCompleteRequest": {
        "properties": {
          "height": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 50000.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "sha256": {
            "anyOf": [
              {
                "pattern": "^[a-fA-F0-9]{64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha256"
          },
          "width": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 50000.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          }
        },
        "title": "UploadCompleteRequest",
        "type": "object"
      },
      "UploadInitiateRequest": {
        "properties": {
          "alt_text": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alt Text"
          },
          "filename": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Filename",
            "type": "string"
          },
          "mime_type": {
            "maxLength": 127,
            "minLength": 3,
            "title": "Mime Type",
            "type": "string"
          },
          "size_bytes": {
            "exclusiveMinimum": 0.0,
            "title": "Size Bytes",
            "type": "integer"
          },
          "visibility": {
            "default": "PUBLIC",
            "enum": [
              "PUBLIC",
              "PRIVATE"
            ],
            "title": "Visibility",
            "type": "string"
          }
        },
        "required": [
          "filename",
          "mime_type",
          "size_bytes"
        ],
        "title": "UploadInitiateRequest",
        "type": "object"
      },
      "UploadInitiateResponse": {
        "properties": {
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "form_fields": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Form Fields",
            "type": "object"
          },
          "media_id": {
            "format": "uuid",
            "title": "Media Id",
            "type": "string"
          },
          "upload_url": {
            "title": "Upload Url",
            "type": "string"
          }
        },
        "required": [
          "media_id",
          "upload_url",
          "form_fields",
          "expires_in"
        ],
        "title": "UploadInitiateResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Domain-resolved, RLS-protected multi-tenant backend.",
    "title": "Kanoon Educational Institute API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/media/uploads": {
      "post": {
        "operationId": "initiate_upload_api_v1_admin_media_uploads_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadInitiateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadInitiateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "413": {
            "$ref": "#/components/responses/Error413"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Initiate Upload",
        "tags": [
          "admin-media"
        ]
      }
    },
    "/api/v1/admin/media/uploads/{media_id}/complete": {
      "post": {
        "operationId": "complete_upload_api_v1_admin_media_uploads__media_id__complete_post",
        "parameters": [
          {
            "in": "path",
            "name": "media_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Media Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadCompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Complete Upload",
        "tags": [
          "admin-media"
        ]
      }
    },
...
    "/api/v1/public/media/{media_id}": {
      "get": {
        "operationId": "download_media_api_v1_public_media__media_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "media_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Media Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "307": {
            "description": "Successful Response"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          }
        },
        "summary": "Download Media",
        "tags": [
          "public-media"
        ]
      }
    },
...
  }
}